feat/panel: Adds output directory and filename selection
Allows the user to specify the output directory and filename for the generated Markdown file.
This enhances flexibility by enabling users to save the output to a location of their choosing and
specify a custom filename.
A default output filename based on the project folder name is set, and the output directory defaults to the current working directory if not specified.